requirement.txt
absl-py==0.7.0
altgraph==0.16.1
asn1crypto==0.24.0
astor==0.7.1
awscli==1.16.106
bcrypt==3.1.6
beautifulsoup4==4.7.1
blessings==1.7
botocore==1.12.96
bpython==0.17.1
certifi==2018.11.29
cffi==1.12.0
chardet==3.0.4
colorama==0.3.9
colored==1.3.93
cryptography==2.5
cssselect==1.0.3
curtsies==0.3.0
cycler==0.10.0
decorator==4.3.2
docutils==0.14
emoji==0.5.1
et-xmlfile==1.0.1
future==0.17.1
gast==0.2.2
greenlet==0.4.15
grpcio==1.18.0
h5py==2.9.0
howdoi==1.1.14
idna==2.8
imageio==2.5.0
IMAPClient==2.1.0
jdcal==1.4
jmespath==0.9.3
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lxml==4.3.1
macholib==1.11
Markdown==3.0.1
matplotlib==3.0.2
mtranslate==1.6
mysql-connector-python==8.0.15
networkx==2.2
numpy==1.16.1
openpyxl==2.6.0
pandas==0.24.1
paramiko==2.4.2
pefile==2018.8.8
Pillow==5.4.1
progressbar==2.5
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.5
PyAutoGUI==0.9.41
pycparser==2.19
PyGetWindow==0.0.4
Pygments==2.3.1
PyInstaller==3.4
PyMsgBox==1.0.6
PyNaCl==1.3.0
pyparsing==2.3.1
PyPDF2==1.26.0
pyperclip==1.7.0
pypiwin32==223
pypyodbc==1.3.4
pyquery==1.4.0
PyRect==0.1.4
PyScreeze==0.1.19
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
python-docx==0.8.10
PyTweening==1.0.3
pytz==2018.9
pywin32==224
pywin32-ctypes==0.2.0
pyxlsb==1.0.4
PyYAML==3.13
pyzmail36==1.0.4
requests==2.21.0
requests-cache==0.4.13
rsa==3.4.2
s3transfer==0.2.0
scipy==1.2.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.7.3
swampy==2.1.7
tensorboard==1.12.2
termcolor==1.1.0
Theano==1.0.4
typing==3.6.6
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
wget==3.2
